moved mock data from individual files to mock data file
