+++ src/routes/media/media-router.int.test.js
@@ -1,0 +1,5 @@
+describe('test', () => {
+  it('should work', () => {
+    expect('test').toBe('test')
+  })
+})
